move components and services into subfolders per style guide
To have cleaner code and follow the style guide
