+++ glassdoor/src/app/app-routing.module.ts
@@ -1,9 +1,9 @@
 import { NgModule }             from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-import { DashboardComponent }   from '../dashboard/dashboard.component';
-import { CompaniesComponent }      from '../companies/companies.component';
-import { CompanyDetailComponent }  from '../company-detail/company-detail.component';
+import { DashboardComponent }   from './dashboard/dashboard.component';
+import { CompaniesComponent }      from './companies/companies.component';
+import { CompanyDetailComponent }  from './companies/company-detail/company-detail.component';
 
 const routes: Routes = [
   { path: '', redirectTo: '/dashboard', pathMatch: 'full' },
