+++ glassdoor/src/app/app.module.ts
@@ -3,18 +3,18 @@ import { NgModule } from '@angular/core';
 import { FormsModule }   from '@angular/forms';
 import { HttpModule }    from '@angular/http';
 
-import { AppRoutingModule } from './app-routing/app-routing.module';
+import { AppRoutingModule } from './app-routing.module';
 
 // Imports for loading & configuring the in-memory web api
 import { InMemoryWebApiModule } from 'angular-in-memory-web-api';
-import { InMemoryDataService }  from './in-memory-data.service';
+import { InMemoryDataService }  from './shared/in-memory-data.service';
 
 import { AppComponent } from './app.component';
 import { CompaniesComponent } from './companies/companies.component';
-import { CompanyDetailComponent } from './company-detail/company-detail.component';
-import { CompanySearchComponent } from './company-search/company-search.component';
+import { CompanyDetailComponent } from './companies/company-detail/company-detail.component';
+import { CompanySearchComponent } from './companies/company-search/company-search.component';
 import { DashboardComponent } from './dashboard/dashboard.component';
-import { CompanyService } from './company.service';
+import { CompanyService } from './companies/shared/company.service';
 
 @NgModule({
   declarations: [
