+++ glassdoor/src/app/companies/company-detail/company-detail.component.ts
@@ -3,8 +3,8 @@ import { Component, OnInit }        from '@angular/core';
 import { ActivatedRoute, ParamMap } from '@angular/router';
 import { Location }                 from '@angular/common';
 
-import { Company }    	from '../company/company';
-import { CompanyService } 	from '../company.service';
+import { Company }    	from '../shared/company.model';
+import { CompanyService } 	from '../shared/company.service';
 
 @Component({
   selector: 'hero-detail',
