+++ glassdoor/src/app/companies/company-search/company-search.component.css
