+++ glassdoor/src/app/companies/company-search/company-search.component.ts
@@ -12,8 +12,8 @@ import 'rxjs/add/operator/catch';
 import 'rxjs/add/operator/debounceTime';
 import 'rxjs/add/operator/distinctUntilChanged';
 
-import { CompanySearchService } from '../company-search.service';
-import { Company } from '../company/company';
+import { CompanySearchService } from '../shared/company-search.service';
+import { Company } from '../shared/company.model';
 
 @Component({
   selector: 'company-search',
