+++ glassdoor/src/app/companies/shared/company-search.service.spec.ts
