+++ glassdoor/src/app/companies/shared/company.model.ts
